+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.21.0] 2026-09-04
+
+### Added
+
+- Add `invoices.paymentSummary` to get the related-document object needed to build a payment complement: installment number, previous balance, and taxes prorated to the paid amount.
+
+### Fixed
+
+- Accept customer ID strings in the `customer` field of `receipts.toInvoice` and `receipts.previewToInvoicePdf` inputs.
+
 ## [4.20.0] 2026-08-21
 
 ### Added